I have cleaned up main.rs by creating a Ui module and Utilis module. I moved the functions from main.rs to their respective modules. I deleted the unused and reduant modules we created at the begining.
