+++ crates/oxidize-mail/src/ui/mod.rs
@@ -1,0 +1,5 @@
+pub mod header_bar;
+pub mod left_pane;
+pub mod middle_pane;
+pub mod right_pane;
+pub mod settings_dialog;
