+++ crates/oxidize-mail/src/utilis/mod.rs
@@ -1,0 +1,3 @@
+pub mod register_resources;
+pub mod style;
+
